Better executable launcher (unarguably)
