+++ ml-agents-envs/mlagents_envs/tests/test_envs.py
@@ -15,7 +15,7 @@ def test_handles_bad_filename(get_communicator):
         UnityEnvironment(" ")
 
 
-@mock.patch("mlagents_envs.environment.UnityEnvironment.executable_launcher")
+@mock.patch("mlagents_envs.env_utils.executable_launcher")
 @mock.patch("mlagents_envs.environment.UnityEnvironment.get_communicator")
 def test_initialization(mock_communicator, mock_launcher):
     mock_communicator.return_value = MockCommunicator(
@@ -37,7 +37,7 @@ def test_initialization(mock_communicator, mock_launcher):
         (None, None, UnityEnvironment.DEFAULT_EDITOR_PORT),
     ],
 )
-@mock.patch("mlagents_envs.environment.UnityEnvironment.executable_launcher")
+@mock.patch("mlagents_envs.env_utils.executable_launcher")
 @mock.patch("mlagents_envs.environment.UnityEnvironment.get_communicator")
 def test_port_defaults(
     mock_communicator, mock_launcher, base_port, file_name, expected
@@ -49,7 +49,7 @@ def test_port_defaults(
     assert expected == env.port
 
 
-@mock.patch("mlagents_envs.environment.UnityEnvironment.executable_launcher")
+@mock.patch("mlagents_envs.env_utils.executable_launcher")
 @mock.patch("mlagents_envs.environment.UnityEnvironment.get_communicator")
 def test_log_file_path_is_set(mock_communicator, mock_launcher):
     mock_communicator.return_value = MockCommunicator()
@@ -61,7 +61,7 @@ def test_log_file_path_is_set(mock_communicator, mock_launcher):
     assert args[log_file_index + 1] == "./some-log-folder-path/Player-0.log"
 
 
-@mock.patch("mlagents_envs.environment.UnityEnvironment.executable_launcher")
+@mock.patch("mlagents_envs.env_utils.executable_launcher")
 @mock.patch("mlagents_envs.environment.UnityEnvironment.get_communicator")
 def test_reset(mock_communicator, mock_launcher):
     mock_communicator.return_value = MockCommunicator(
@@ -84,7 +84,7 @@ def test_reset(mock_communicator, mock_launcher):
         assert (n_agents,) + shape == obs.shape
 
 
-@mock.patch("mlagents_envs.environment.UnityEnvironment.executable_launcher")
+@mock.patch("mlagents_envs.env_utils.executable_launcher")
 @mock.patch("mlagents_envs.environment.UnityEnvironment.get_communicator")
 def test_step(mock_communicator, mock_launcher):
     mock_communicator.return_value = MockCommunicator(
@@ -122,7 +122,7 @@ def test_step(mock_communicator, mock_launcher):
     assert 2 in terminal_steps
 
 
-@mock.patch("mlagents_envs.environment.UnityEnvironment.executable_launcher")
+@mock.patch("mlagents_envs.env_utils.executable_launcher")
 @mock.patch("mlagents_envs.environment.UnityEnvironment.get_communicator")
 def test_close(mock_communicator, mock_launcher):
     comm = MockCommunicator(discrete_action=False, visual_inputs=0)
